Added NavBar and Footer
A styled navbar with logo, cart and two product sections. Future search bar integration is in the works. A simple footer section was also added. Many styling changes to the homepage were done. Installed tailwindcss for future styling, but it did not work properly, must troubleshoot at a later point.